feat: add `release.openRelease` task
