+++ lib/index.js
@@ -3,7 +3,6 @@ const args = require('args')
 const chalk = require('chalk')
 const semVer = require('semver')
 const inquirer = require('inquirer')
-const open = require('open')
 const taggedVersions = require('tagged-versions')
 
 // Ours
@@ -42,16 +41,7 @@ const changeTypes = [
   }
 ]
 
-const getReleaseURL = (release, edit = false) => {
-  if (!release || !release.html_url) {
-    return false
-  }
-
-  const htmlURL = release.html_url
-  return edit ? htmlURL.replace('/tag/', '/edit/') : htmlURL
-}
-
-const createRelease = (changelog, options) => {
+const createRelease = async (changelog, options) => {
   const {tag, github, repo: {user: owner, repo}, release: {id} = {}} = options
   const isPre = flags.pre ? 'pre' : ''
   handleSpinner.create(`Uploading ${isPre}release`)
@@ -73,21 +63,14 @@ const createRelease = (changelog, options) => {
     body.id = id
   }
 
-  github.repos[method](body, (err, response) => {
-    if (err) {
-      console.log('\n')
-      handleSpinner.fail('Failed to upload release.')
-    }
+  const release = await github.repos[method](body).catch(
+    () => Promise.reject(new Error('Failed to upload release.'))
+  )
 
-    global.spinner.succeed()
-    const releaseURL = getReleaseURL(response, true)
+  global.spinner.succeed()
+  console.log(`\n${chalk.bold('Done!')} 🎉 Opening release in browser...`)
 
-    if (releaseURL) {
-      open(releaseURL)
-    }
-
-    console.log(`\n${chalk.bold('Done!')} 🎉 Opening release in browser...`)
-  })
+  tasks.openRelease({release, edit: true})
 }
 
 const orderCommits = (commits, options) => {
@@ -187,16 +170,9 @@ module.exports = exports = async () => {
   global.spinner.succeed()
   console.log('')
 
-  const releaseURL = getReleaseURL(release)
+  tasks.openRelease({release})
 
-  if (releaseURL) {
-    open(releaseURL)
-  }
-
-  const alreadyThere = 'Release already exists. Opening in browser...'
-  console.error(`${chalk.red('Error!')} ` + alreadyThere)
-
-  process.exit(1)
+  throw new Error('Release already exists. Opening in browser...')
 }
 
 Object.assign(exports, tasks, {tap})
